Finish route tests

+++ api/businesses.js
@@ -43,71 +43,63 @@ router.get('/', async function (req, res, next) {
     const db = MongoDB.getInstance();
     const businessesColl = db.collection("businesses");
     const query = {};
-    const isAdmin = await checkAdmin(req)
-    if (isAdmin) {
-        try {
-            const cursor = businessesColl.find(query);
-            await cursor.forEach(doc => {
-                businesses.push(doc);
-            });
-
-            console.dir(businesses);
-        } catch (error) {
-            console.error("Error:", error);
-            res.status(400).json({
-                error: error
-            });
-        }
-
-        /*
-         * Compute page number based on optional query string parameter `page`.
-         * Make sure page is within allowed bounds.
-         */
-
-        let page = parseInt(req.query.page) || 1;
-        const numPerPage = 10;
-        const lastPage = Math.ceil(businesses.length / numPerPage);
-        page = page > lastPage ? lastPage : page;
-        page = page < 1 ? 1 : page;
-
-        /*
-         * Calculate starting and ending indices of businesses on requested page and
-         * slice out the corresponsing sub-array of busibesses.
-         */
-        const start = (page - 1) * numPerPage;
-        const end = start + numPerPage;
-        const pageBusinesses = businesses.slice(start, end);
-
-        /*
-         * Generate HATEOAS links for surrounding pages.
-         */
-        const links = {};
-        if (page < lastPage) {
-            links.nextPage = `/businesses?page=${page + 1}`;
-            links.lastPage = `/businesses?page=${lastPage}`;
-        }
-        if (page > 1) {
-            links.prevPage = `/businesses?page=${page - 1}`;
-            links.firstPage = '/businesses?page=1';
-        }
-
-        /*
-         * Construct and send response.
-         */
-        res.status(200).json({
-            businesses: pageBusinesses,
-            pageNumber: page,
-            totalPages: lastPage,
-            pageSize: numPerPage,
-            totalCount: businesses.length,
-            links: links
+    try {
+        const cursor = businessesColl.find(query);
+        await cursor.forEach(doc => {
+            businesses.push(doc);
         });
-    } else {
-        res.status(403).json({
-            error: "Unauthorized to access the specified resource"
+
+        console.dir(businesses);
+    } catch (error) {
+        console.error("Error:", error);
+        res.status(400).json({
+            error: error
         });
     }
 
+    /*
+     * Compute page number based on optional query string parameter `page`.
+     * Make sure page is within allowed bounds.
+     */
+
+    let page = parseInt(req.query.page) || 1;
+    const numPerPage = 10;
+    const lastPage = Math.ceil(businesses.length / numPerPage);
+    page = page > lastPage ? lastPage : page;
+    page = page < 1 ? 1 : page;
+
+    /*
+     * Calculate starting and ending indices of businesses on requested page and
+     * slice out the corresponsing sub-array of busibesses.
+     */
+    const start = (page - 1) * numPerPage;
+    const end = start + numPerPage;
+    const pageBusinesses = businesses.slice(start, end);
+
+    /*
+     * Generate HATEOAS links for surrounding pages.
+     */
+    const links = {};
+    if (page < lastPage) {
+        links.nextPage = `/businesses?page=${page + 1}`;
+        links.lastPage = `/businesses?page=${lastPage}`;
+    }
+    if (page > 1) {
+        links.prevPage = `/businesses?page=${page - 1}`;
+        links.firstPage = '/businesses?page=1';
+    }
+
+    /*
+     * Construct and send response.
+     */
+    res.status(200).json({
+        businesses: pageBusinesses,
+        pageNumber: page,
+        totalPages: lastPage,
+        pageSize: numPerPage,
+        totalCount: businesses.length,
+        links: links
+    });
 });
 
 /*
@@ -115,9 +107,10 @@ router.get('/', async function (req, res, next) {
  */
 router.post('/', requireAuthentication, async function (req, res, next) {
 
+    const isAdmin = await checkAdmin(req);
     if (validateAgainstSchema(req.body, businessSchema)) {
         const business = extractValidFields(req.body, businessSchema);
-        if (req.user !== business.ownerid) {
+        if (!isAdmin && req.user !== business.ownerid) {
             res.status(403).json({
                 error: "Unauthorized to access the specified resource"
             });
@@ -192,7 +185,8 @@ router.get('/:businessid', async function (req, res, next) {
 /*
  * Route to replace data for a business.
  */
-router.put('/:businessid', async function (req, res, next) {
+router.put('/:businessid', requireAuthentication, async function (req, res, next) {
+
     let businessid;
     try {
         businessid = ObjectId.createFromHexString(req.params.businessid);
@@ -211,24 +205,32 @@ router.put('/:businessid', async function (req, res, next) {
         console.log("No businesses found");
         next();
     }
+
+    const isAdmin = await checkAdmin(req);
     if (validateAgainstSchema(req.body, businessSchema)) {
         const updatedBusiness = extractValidFields(req.body, businessSchema);
-        try {
-            const result = await businessesColl.replaceOne(query, updatedBusiness);
-            if (result.matchedCount === 1 && result.modifiedCount === 1) {
-                res.status(200).json({
-                    links: {
-                        business: `/businesses/${businessid}`
-                    }
-                });
-            } else {
-                next();
-            }
-        } catch (error) {
-            console.error("Error:", error);
-            res.status(400).json({
-                error: error
+        if (!isAdmin && req.user !== updatedBusiness.ownerid) {
+            res.status(403).json({
+                error: "Unauthorized to access the specified resource"
             });
+        } else {
+            try {
+                const result = await businessesColl.replaceOne(query, updatedBusiness);
+                if (result.matchedCount === 1 && result.modifiedCount === 1) {
+                    res.status(200).json({
+                        links: {
+                            business: `/businesses/${businessid}`
+                        }
+                    });
+                } else {
+                    next();
+                }
+            } catch (error) {
+                console.error("Error:", error);
+                res.status(400).json({
+                    error: error
+                });
+            }
         }
     } else {
         res.status(400).json({
@@ -240,7 +242,7 @@ router.put('/:businessid', async function (req, res, next) {
 /*
  * Route to delete a business.
  */
-router.delete('/:businessid', async function (req, res, next) {
+router.delete('/:businessid', requireAuthentication, async function (req, res, next) {
 
     let businessid;
     try {
@@ -261,18 +263,34 @@ router.delete('/:businessid', async function (req, res, next) {
         next();
     }
 
+    let existingBusiness;
     try {
-        const result = await businessesColl.deleteOne(query);
-        if (result.deletedCount === 1) {
-            res.status(204).end();
-        }
-        else {
-            next();
-        }
+        existingBusiness = await businessesColl.findOne(query);
+        if (!existingBusiness) throw new Error("Business not found");
     } catch (error) {
-        console.error("Error:", error);
-        res.status(400).json({
+        return res.status(404).json({
             error: error
+        })
+    }
+    const isAdmin = await checkAdmin(req);
+    if (!isAdmin && req.user !== existingBusiness.ownerid) {
+        res.status(403).json({
+            error: "Unauthorized to access the specified resource"
         });
+    } else {
+        try {
+            const result = await businessesColl.deleteOne(query);
+            if (result.deletedCount === 1) {
+                res.status(204).end();
+            }
+            else {
+                next();
+            }
+        } catch (error) {
+            console.error("Error:", error);
+            res.status(400).json({
+                error: error
+            });
+        }
     }
 });
