+++ api/photos.js
@@ -3,6 +3,7 @@ const { validateAgainstSchema, extractValidFields } = require('../lib/validation
 
 const MongoDB = require('../database');
 const { ObjectId } = require('mongodb');
+const { requireAuthentication, checkAdmin } = require('../lib/auth');
 const photoCollection = "photos";
 
 exports.photoCollection = photoCollection;
@@ -21,19 +22,27 @@ const photoSchema = {
 /*
  * Route to create a new photo.
  */
-router.post('/', async function (req, res, next) {
+router.post('/', requireAuthentication, async function (req, res, next) {
+
+    const isAdmin = await checkAdmin(req);
     if (validateAgainstSchema(req.body, photoSchema)) {
         const photo = extractValidFields(req.body, photoSchema);
-        const db = MongoDB.getInstance();
-        const photoColl = db.collection(photoCollection);
-        const result = await photoColl.insertOne(photo);
-        res.status(201).json({
-            id: result.insertedId,
-            links: {
-                photo: `/photos/${result.insertedId}`,
-                business: `/businesses/${photo.businessid}`
-            }
-        });
+        if (!isAdmin && req.user !== photo.userid) {
+            res.status(403).json({
+                error: "Unauthorized to access the specified resource"
+            });
+        } else {
+            const db = MongoDB.getInstance();
+            const photoColl = db.collection(photoCollection);
+            const result = await photoColl.insertOne(photo);
+            res.status(201).json({
+                id: result.insertedId,
+                links: {
+                    photo: `/photos/${result.insertedId}`,
+                    business: `/businesses/${photo.businessid}`
+                }
+            });
+        }
     } else {
         res.status(400).json({
             error: "Request body is not a valid photo object"
@@ -82,7 +91,7 @@ router.get('/:photoID', async function (req, res, next) {
 /*
  * Route to update a photo.
  */
-router.put('/:photoID', async function (req, res, next) {
+router.put('/:photoID', requireAuthentication, async function (req, res, next) {
 
     let photoID;
     try {
@@ -102,36 +111,44 @@ router.put('/:photoID', async function (req, res, next) {
         console.log("No photos found");
         next();
     }
+
+    const isAdmin = await checkAdmin(req);
     if (validateAgainstSchema(req.body, photoSchema)) {
         /*
          * Make sure the updated photo has the same businessid and userid as
          * the existing photo.
          */
         const updatedPhoto = extractValidFields(req.body, photoSchema);
-        try {
-            const existingPhoto = await photoColl.findOne(query);
-            if (existingPhoto && updatedPhoto.businessid === existingPhoto.businessid && updatedPhoto.userid === existingPhoto.userid) {
-                const result = await photoColl.replaceOne(query, updatedPhoto);
-                if (result.matchedCount === 1 && result.modifiedCount === 1) {
-                    res.status(200).json({
-                        links: {
-                            photo: `/photos/${photoID}`,
-                            business: `/businesses/${updatedPhoto.businessid}`
-                        }
-                    });
+        if (!isAdmin && req.user !== updatedPhoto.userid) {
+            res.status(403).json({
+                error: "Unauthorized to access the specified resource"
+            });
+        } else {
+            try {
+                const existingPhoto = await photoColl.findOne(query);
+                if (existingPhoto && updatedPhoto.businessid === existingPhoto.businessid && updatedPhoto.userid === existingPhoto.userid) {
+                    const result = await photoColl.replaceOne(query, updatedPhoto);
+                    if (result.matchedCount === 1 && result.modifiedCount === 1) {
+                        res.status(200).json({
+                            links: {
+                                photo: `/photos/${photoID}`,
+                                business: `/businesses/${updatedPhoto.businessid}`
+                            }
+                        });
+                    } else {
+                        next();
+                    }
                 } else {
-                    next();
+                    res.status(403).json({
+                        error: "Updated photo cannot modify businessid or userid"
+                    });
                 }
-            } else {
-                res.status(403).json({
-                    error: "Updated photo cannot modify businessid or userid"
+            } catch (error) {
+                console.error("Error:", error);
+                res.status(400).json({
+                    error: error
                 });
             }
-        } catch (error) {
-            console.error("Error:", error);
-            res.status(400).json({
-                error: error
-            });
         }
     } else {
         res.status(400).json({
@@ -143,7 +160,7 @@ router.put('/:photoID', async function (req, res, next) {
 /*
  * Route to delete a photo.
  */
-router.delete('/:photoID', async function (req, res, next) {
+router.delete('/:photoID', requireAuthentication, async function (req, res, next) {
 
     let photoID;
     try {
@@ -163,15 +180,30 @@ router.delete('/:photoID', async function (req, res, next) {
         console.log("No photos found");
         next();
     }
-
+    let existingPhoto;
     try {
-        const result = await photoColl.deleteOne(query);
-        if (result.deletedCount === 1) {
-            res.status(204).end();
-        } else {
-            next();
-        }
+        existingPhoto = await photoColl.findOne(query);
+        if (!existingPhoto) throw new Error("Photo not found");
     } catch (error) {
-        console.error("Error:", error);
+        return res.status(400).json({
+            error: error
+        })
+    }
+    const isAdmin = await checkAdmin(req);
+    if (!isAdmin && req.user !== existingPhoto.userid) {
+        res.status(403).json({
+            error: "Unauthorized to access the specified resource"
+        });
+    } else {
+        try {
+            const result = await photoColl.deleteOne(query);
+            if (result.deletedCount === 1) {
+                res.status(204).end();
+            } else {
+                next();
+            }
+        } catch (error) {
+            console.error("Error:", error);
+        }
     }
 });
